created sub-component for question fields (second column)
